Changed the background, Refactoring of code.
