+++ src/Components/mint/Mint.jsx
@@ -5,14 +5,19 @@ import { ethers } from "ethers";
 import { useForm } from "react-hook-form";
 
 import { ContractABI } from "./contract";
+const background = require("../../images/landing/Herobackground.png");
 
-function Mint() {
-  const [wallet, setWallet] = useState("Connect a Wallet");
-  const [logout, setLogout] = useState(false);
-  const [maxMintAmount, setMaxMintAmount] = useState();
-  const [price, setPrice] = useState(0);
-  const [images, setImages] = useState([]);
-  const [userMintedAmount, setUserMintedAmount] = useState(0);
+function Mint({
+  wallet,
+  price,
+  images,
+  userMintedAmount,
+  maxMintAmount,
+  disconnect,
+  connection,
+  readContract,
+  getTokens,
+}) {
   const [check, setCheck] = useState(false);
   const customId = "custom-id-yes";
   const notify = (message) => {
@@ -21,9 +26,7 @@ function Mint() {
     });
   };
 
-  const { REACT_APP_NETWORK } = process.env;
   const { REACT_APP_CONTRACT_ADDRESS } = process.env;
-  const { REACT_APP_NETWORK_CHAIN_ID } = process.env;
   const { register, handleSubmit } = useForm();
 
   const onSubmit = async (data) => {
@@ -33,47 +36,6 @@ function Mint() {
     await getTokens();
   };
 
-  const setupConnections = async () => {
-    if (window.ethereum != null) {
-      const provider = new ethers.providers.Web3Provider(window.ethereum);
-      const network = await provider.getNetwork();
-      if (REACT_APP_NETWORK !== network.name) {
-        notify(
-          `Not on a correct network. Change your network to "${REACT_APP_NETWORK}"`
-        );
-        return false;
-      } else {
-        await provider.send("eth_requestAccounts", []);
-        const signer = await provider.getSigner();
-        const address = await signer.getAddress();
-        return address;
-      }
-    } else {
-      notify("No Ether wallet available");
-      return false;
-    }
-  };
-  const connection = async () => {
-    const res = await setupConnections();
-    if (res === false) {
-      setWallet("Connect a Wallet");
-      await window.ethereum.request({
-        method: "wallet_switchEthereumChain",
-        params: [{ chainId: REACT_APP_NETWORK_CHAIN_ID }],
-      });
-    } else {
-      setLogout(true);
-      setWallet(res.slice(0, 6) + "..." + res.slice(36, 42));
-    }
-  };
-  const disconnect = async () => {
-    setWallet("Connect a Wallet");
-    setLogout(false);
-    setUserMintedAmount("-");
-    setMaxMintAmount("-");
-    setPrice("-");
-    setImages([]);
-  };
   const mint = async (mintAmount) => {
     setCheck(!check);
     if (wallet === "Connect a Wallet") {
@@ -103,45 +65,6 @@ function Mint() {
       }
     }
   };
-  const getTokens = async () => {
-    const provider = new ethers.providers.Web3Provider(window.ethereum);
-    const contract = new ethers.Contract(
-      REACT_APP_CONTRACT_ADDRESS,
-      ContractABI,
-      provider
-    );
-    let accounts = await provider.send("eth_requestAccounts", []);
-    let address = accounts[0];
-    const imagesTockens = await contract.nftsOnwedByWallet(address);
-    let imagesLocal = [];
-    await imagesTockens.map(async (image) => {
-      const url = await contract.tokenURI(parseInt(image, 10));
-      let result = await url.replace("ipfs://", "https://ipfs.io/ipfs/");
-      const jsonBody = await (await fetch(result)).json();
-      imagesLocal.push(
-        await jsonBody.image.replace("ipfs://", "https://ipfs.io/ipfs/")
-      );
-    });
-    setTimeout(() => {
-      setImages(imagesLocal);
-    }, [1000]);
-  };
-  const readContract = async () => {
-    const provider = new ethers.providers.Web3Provider(window.ethereum);
-    const contract = new ethers.Contract(
-      REACT_APP_CONTRACT_ADDRESS,
-      ContractABI,
-      provider
-    );
-    const maxMintAmount = await contract.maxMintAmount();
-    let accounts = await provider.send("eth_requestAccounts", []);
-    let address = accounts[0];
-    const userMintedAmount = await contract.balanceOf(address);
-    const price = await contract.cost();
-    setMaxMintAmount(parseInt(maxMintAmount, 10));
-    setUserMintedAmount(parseInt(userMintedAmount, 10));
-    setPrice(Number(ethers.utils.formatEther(price)));
-  };
 
   useEffect(() => {
     if (window.ethereum) {
@@ -156,77 +79,60 @@ function Mint() {
       });
     }
   });
-  const fun = async () => {
+  const initialFun = async () => {
     await connection();
     await readContract();
     await getTokens();
   };
   useEffect(() => {
-    fun();
+    initialFun();
   }, []);
 
   return (
     <>
-      <div className="bg-dark">
-        <div className="text-white flex flex-wrap  justify-between md:px-36 px-10 py-8">
-          <div className="text-3xl">Dashboard</div>
-          <div>
-            <button
-              className="text-white w-[200px] shadow-lg border items-center text-xl bg-gray-700 rounded-xl"
-              onClick={() => {
-                connection();
-                readContract();
-                getTokens();
-              }}
-            >
-              {wallet}
-            </button>
-            <button
-              className="grid mt-3 justify-center w-[200px] shadow-lg border items-center text-white text-xl bg-gray-700 rounded-xl"
-              onClick={() => {
-                disconnect();
-                setUserMintedAmount("-");
-                setMaxMintAmount("-");
-                setPrice("-");
-                setImages([]);
-              }}
-              style={{ display: logout ? "block" : "none" }}
-            >
-              Disconnect
-            </button>
+      <div className="top-[10%] -left-[0px] z-20 lg:h-screen h-full w-full absolute object-center object-cover hidden md:block">
+        <img src={background} alt="/" />
+      </div>
+      <div className=" z-30 pt-28 pb-[0rem] relative">
+        <div className="flex justify-center">
+          <div className="bg-slate-300 bg-opacity-80 font-poppins rounded-[55px] px-20 py-10">
+            <div className="flex justify-end ">
+              <form
+                onSubmit={handleSubmit(onSubmit)}
+                className="flex justify-center"
+              >
+                <div class="">
+                  <input
+                    type="number"
+                    className=" text-black rounded-full focus:outline-none bg-white mt-2 p-3"
+                    {...register("value", { required: true })}
+                    defaultValue="0"
+                  />
+                </div>
+                <button className="text-white h-[45px] mt-2 bg-gradient-to-br from-pink-500 to-orange-400 hover:bg-gradient-to-bl focus:ring-4 focus:outline-none focus:ring-pink-200 dark:focus:ring-pink-800 font-medium text-md rounded-[55px] px-5 mx-5  text-center  duration-700 hover:scale-110">
+                  Mint
+                </button>
+              </form>
+            </div>
+            <h4 className="text-white text-center pt-5 font-bold text-3xl">
+              Total Minted: {userMintedAmount}
+            </h4>
+            <h4 className="text-white text-center font-bold text-3xl">
+              Maximum Minted: {maxMintAmount}
+            </h4>
+            <h4 className="text-white text-center font-bold text-3xl">
+              Price: {price} eth
+            </h4>
+            <h4 className="text-white text-center font-bold text-3xl">
+              Remaining minting:{maxMintAmount - userMintedAmount}
+            </h4>
           </div>
         </div>
-
-        <div className="flex flex-wrap justify-center text-white">
-          <form onSubmit={handleSubmit(onSubmit)}>
-            <input
-              type="number"
-              className="rounded px-2 py-3 text-black"
-              {...register("value", { required: true })}
-              defaultValue="0"
-            />
-            <button className="border rounded-lg text-2xl px-5 ml-5 py-1 my-5 bg-gray-500">
-              Mint
-            </button>
-          </form>
-        </div>
-        <h4 className="text-white text-center font-bold text-3xl">
-          Total Minted: {userMintedAmount}
-        </h4>
-        <h4 className="text-white text-center font-bold text-3xl">
-          Maximum Minted: {maxMintAmount}
-        </h4>
-        <h4 className="text-white text-center font-bold text-3xl">
-          Price: {price} eth
-        </h4>
-        <h4 className="text-white text-center font-bold text-3xl">
-          Remaining minting:{maxMintAmount - userMintedAmount}
-        </h4>
         <div className="text-white justify-center px-10 py-10 grid text-center lg:grid-cols-3 md:grid-cols-3 sm:grid-cols-2 grid-cols-1">
           {images.map((url, key) => {
             return (
               <div key={key} className="border border-md">
-                <img alt="Tokens images" src={url} />
+                <img key={key} alt="Tokens images" src={url} />
               </div>
             );
           })}
